+++ CHANGELOG.md
@@ -28,7 +28,7 @@
     * Added functional test for -Process Template option.
     * Added examples and parameter description to help.
 1. For WorkItemType operations, **added function** a private/helper `_getProcessTemplateUrl` in file `common.ps1` to get/set the urls of templates in a script level hash table (to reduce API calls to get ID from a name.) In a previous version I set a URL property on the processs object - this has been dropped.
-1. **Added functions**  `Add-VSTeamWorkItemType`, `Set-VSTeamWorkItemType`, `Remove-VSTeamWorkItemType` with associated help and tests.
+1. **Added functions**  `Add-VSTeamWorkItemType`, `Set-VSTeamWorkItemType`, `Remove-VSTeamWorkItemType`, `Unlock-VSTeamWorkItemType` with associated help and tests.
 1. Ensured tests which mock Get-VSTeamProcess clear the process cache (otherwise the processes they return will not be used to validate the Process parameter during the test,)
 1. Ensured tests which call Set-VSTeamProject load and mock Get-VSTeamProcess.